Switched back to renderer.js
