+++ Screenshot-App/index.html
@@ -16,7 +16,7 @@
 		<button id='screen-shot' class = "btn btn-primary">Take Screenshot</button>
 	</div>
 	<script>if (typeof module === 'object'){window.module = module; module = undefined;}</script>
-	<script type="text/javascript" src="renderer1.js"></script>
+	<script type="text/javascript" src="renderer.js"></script>
 	<script>if (window.module) module = window.module;</script>
 
 </body>
